us the transition matrix to calculate the probability of default and survival directly rather than from the calculating the cumulative and then marginal probabilities.
The full TM is now passed through the Z framework.

The probability of cure and write-off now need to be configured in the TTC TM. (see example)

Ratings now start at 1 rather than 0. The 0 rating is used for attrition.

The model can now cater for single and multiple default definitions via the TM approach.